V0.1.5.1
Need to fix variable jump height and do the feedback.

Week 4 Formative Feedback

C1: Make sure dead code is removed for the working implementation
C2: Try to get one more mechanic, regardless if it is a bit buggy.
C3: Decent, make sure to have the mechanics to test on it.
C4: Only thing to add to the formatives would be a short youtube video / gif to demo straight how it works.